fix: mpstat cpu usage values
Fixed issue where mpstat was not returning the correct values. Previously mpstat was being called each time the screen was being updated, but that is not how mpstat is intended to be used. It was returning the cpu usage overall since (I think) the last boot. Now mpstat is being called from within a goroutine and it is correctly returning average cpu usage for each second of monitoring.